idp_manuscript/requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
pytesseract>=0.3.10
PyYAML>=6.0
scikit-image>=0.21.0
matplotlib>=3.7.0
